big change, new theme v1 working
will still want to create a custom theme I would expect... fork the other one, fix that .Hugo.Generator / hugo.generator bug,  then pull in new updates to it as a side gig.

The changes in this commit are around layout, and basically stuffing the content folder into this repo for now until we can break it back out as a submodule and include it at build time dynamicaly.